:bookmark: Update version to 1.8.3 (:pencil2: Re: can you take a look at https://github.com/FCP-INDI/C-PAC/pull/1659/files?short_path=06572a9#diff-06572a96a58dc510037d5efa622f9bec8519bc1beab13c9f251e97e657a9d4ed
so over here the first point is correct that if the user wants to warp
tissue mask to EPI template then EPI_registration under
functional_registration should be On and the template_for_segmentation
should be EPI_template.  The second point is in order to warp tissue
mask to T1 template for this anatomical_registration should be On and
the template_for_segmentation should be T1_template.

[skip ci])